requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.0.0  # For DashScope and OpenAI API compatibility

# Vector databases
chromadb>=0.4.0
faiss-cpu>=1.7.4

# Scientific data processing
biopython>=1.81
numpy>=1.24.0
pandas>=2.0.0

# HTTP and async
requests>=2.31.0
aiohttp>=3.8.0

# Text processing
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0

# Optional: Web interface
# streamlit>=1.28.0
# fastapi>=0.100.0
# uvicorn>=0.23.0